+++ src/LASRcore/Stage.h
@@ -32,6 +32,46 @@
 // JSON parser
 #include "nlohmann/json.hpp"
 
+/* The order the functions are called
+ *
+ * In Pipeline::parser() to build the pipeline
+ *  1. Constructor
+ *  2. set_uid()
+ *  3. set_ncpu()
+ *  4. set_verbose()
+ *  5. set_extent()
+ *  6. connect()
+ *  7. set_parameters()
+ *  8. set_crs()
+ *  9. get_crs()
+ *  10. set_filter()
+ *  11. set_output_file()
+ * In the copy constructor of a Pipeline
+ *  12. clone()
+ * In Pipeline::set_chunk()
+ *  13. set_chunk()
+ *  14. break_pipeline()
+ *  15. set_input_file_name()
+ * In Pipeline::run()
+ *  16. break_pipeline()
+ *  17. process()
+ *  18. process(LASheader)
+ *  19. set_header()
+ *  20. process(LAS)
+ *  21. write()
+ *  22. reset_filter()
+ *  23. clear()
+ *  24. clean()
+ * Pipeline::clear(true)
+ *  23. clear(true)
+ * Pipeline::merge()
+ *  26. merge()
+ * Pipeline::sort()
+ *  27. sort()
+ * Pipeline::to_R()
+ *  28. to_R();
+ */
+
 class Stage
 {
 public:
@@ -46,23 +86,28 @@ public:
   virtual bool break_pipeline() { return false; };
   virtual bool write() { return true; };
   virtual void clear(bool last = false) { return; };
-  virtual bool set_chunk(const Chunk& chunk);
   virtual void set_crs(const CRS& crs) { this->crs = crs; };
   virtual bool set_output_file(const std::string& file) { ofile = file; return true; };
   virtual bool set_input_file_name(const std::string& file) { return true; };
   virtual void set_header(LASheader*& header) { return; };
+  virtual bool set_chunk(Chunk& chunk);
+  virtual bool set_parameters(const nlohmann::json&) { return true; };
   virtual bool is_streamable() const { return false; };
   virtual bool is_parallelizable() const { return true; }; // concurrent-files
   virtual bool is_parallelized() const { return false; };  // concurrent-points
   virtual bool use_rcapi() const { return false; };
   virtual double need_buffer() const { return 0; };
   virtual bool need_points() const { return true; };
-  virtual bool set_parameters(const nlohmann::json&) { return true; };
+  virtual void get_extent(double& xmin, double& ymin, double& xmax, double& ymax) { return; };
+
+
   virtual bool connect(const std::list<std::unique_ptr<Stage>>&, const std::string& uid) { return true; };
+
   //virtual void convert_units() { return; };
 
   virtual std::string get_name() const = 0;
 
+  void set_extent(double xmin, double ymin, double xmax, double ymax) { this->xmin = xmin; this->ymin = ymin; this->xmax = xmax; this->ymax = ymax; };
   void set_ncpu(int ncpu) { this->ncpu = ncpu; }
   void set_ncpu_concurrent_files(int ncpu) { this->ncpu_concurrent_files = ncpu; }
   void set_verbose(bool verbose) { this->verbose = verbose; };
@@ -70,7 +115,6 @@ public:
   void set_filter(const std::string& f);
   void set_progress(Progress* progress) { this->progress = progress; };
   void set_chunk(double xmin, double ymin, double xmax, double ymax) { this->xmin = xmin; this->ymin = ymin; this->xmax = xmax; this->ymax = ymax; };
-  void set_extent(double xmin, double ymin, double xmax, double ymax) { this->xmin = xmin; this->ymin = ymin; this->xmax = xmax; this->ymax = ymax; };
   void reset_filter() { lasfilter.reset(); };
 
   std::string get_uid() const { return uid; };
@@ -153,7 +197,7 @@ public:
   StageRaster();
   StageRaster(const StageRaster& other);
   ~StageRaster() override;
-  bool set_chunk(const Chunk& chunk) override;
+  bool set_chunk(Chunk& chunk) override;
   void set_crs(const CRS& crs) override;
   bool set_input_file_name(const std::string& file) override;
   bool set_output_file(const std::string& file) override;
@@ -171,7 +215,7 @@ public:
   StageVector();
   StageVector(const StageVector& other);
   ~StageVector() override;
-  bool set_chunk(const Chunk& chunk) override;
+  bool set_chunk(Chunk& chunk) override;
   void set_crs(const CRS& crs) override;
   bool set_input_file_name(const std::string& file) override;
   bool set_output_file(const std::string& file) override;
@@ -180,6 +224,24 @@ public:
 
 protected:
   Vector vector;
+};
+
+class StageMatrix : public Stage
+{
+public:
+  StageMatrix();
+  StageMatrix(const StageMatrix& other);
+  void merge(const Stage* other) override;
+  void transform(double& x, double& y, double& z);
+
+  #ifdef USING_R
+    SEXP to_R() override;
+  #endif
+
+  nlohmann::json to_json() const override;
+
+protected:
+  double matrix[4][4];
 };
 
 template <typename T>