+++ src/LASRstages/transformwith.h
@@ -3,8 +3,6 @@
 
 #include "Stage.h"
 
-class LASRtriangulate;
-
 class LASRtransformwith: public Stage
 {
 public:
@@ -14,6 +12,8 @@ public:
   bool set_parameters(const nlohmann::json&) override;
   std::string get_name() const override { return "transform_with"; }
   bool is_parallelized() const override { return true; };
+  bool set_chunk(Chunk& chunk) override;
+  void get_extent(double& xmin, double& ymin, double& xmax, double& ymax) override;
 
   // multi-threading
   LASRtransformwith* clone() const override { return new LASRtransformwith(*this); };